requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
djangorestframework==3.13.1
pip==19.2.3
pytz==2022.1
setuptools==41.2.0
sqlparse==0.4.2
tzdata==2022.1
python-dotenv==0.14.0
psycopg2
PyJWT==1.7.1
requests==2.23.0